src/pkg/file_preview/requirements.txt
# Dependencies for the file preview package
Pillow>=9.0.0  # Image processing library